ixbrl-viewer localviewer: allow runner code to be shared; write viewer to temp dir
